HV-1526 Add different implementations of ValidationContext
- add implementations for bean, property/value, executable parameter
and executable return value contexts
- extract builder for validation contexts to its own class
