+++ engine/src/main/java/org/hibernate/validator/internal/engine/validationcontext/BeanValidationContext.java
@@ -1,0 +1,77 @@
+/*
+ * Hibernate Validator, declare and validate application constraints
+ *
+ * License: Apache License, Version 2.0
+ * See the license.txt file in the root directory or <http://www.apache.org/licenses/LICENSE-2.0>.
+ */
+package org.hibernate.validator.internal.engine.validationcontext;
+
+import javax.validation.ConstraintValidatorFactory;
+import javax.validation.ConstraintViolation;
+import javax.validation.Path;
+import javax.validation.TraversableResolver;
+import javax.validation.metadata.ConstraintDescriptor;
+
+import org.hibernate.validator.constraintvalidation.HibernateConstraintValidatorInitializationContext;
+import org.hibernate.validator.internal.engine.ConstraintViolationImpl;
+import org.hibernate.validator.internal.engine.ValueContext;
+import org.hibernate.validator.internal.engine.constraintvalidation.ConstraintValidatorManager;
+import org.hibernate.validator.internal.engine.constraintvalidation.ConstraintViolationCreationContext;
+import org.hibernate.validator.internal.metadata.aggregated.BeanMetaData;
+
+/**
+ * Implementation of {@link AbstractValidationContext} for bean validation.
+ *
+ * @author Marko Bekhta
+ */
+class BeanValidationContext<T> extends AbstractValidationContext<T> {
+
+	BeanValidationContext(
+			ConstraintValidatorManager constraintValidatorManager,
+			ConstraintValidatorFactory constraintValidatorFactory,
+			ValidatorScopedContext validatorScopedContext,
+			TraversableResolver traversableResolver,
+			HibernateConstraintValidatorInitializationContext constraintValidatorInitializationContext,
+			T rootBean,
+			Class<T> rootBeanClass,
+			BeanMetaData<T> rootBeanMetaData
+	) {
+		super( constraintValidatorManager, constraintValidatorFactory, validatorScopedContext, traversableResolver, constraintValidatorInitializationContext,
+				rootBean, rootBeanClass, rootBeanMetaData, buildDisableAlreadyValidatedBeanTracking( rootBeanMetaData )
+		);
+	}
+
+	private static boolean buildDisableAlreadyValidatedBeanTracking(BeanMetaData<?> rootBeanMetaData) {
+		return !rootBeanMetaData.hasCascadables();
+	}
+
+	@Override
+	protected ConstraintViolation<T> createConstraintViolation(
+			String messageTemplate, String interpolatedMessage, Path propertyPath,
+			ConstraintDescriptor constraintDescriptor, ValueContext localContext,
+			ConstraintViolationCreationContext constraintViolationCreationContext) {
+		return ConstraintViolationImpl.forBeanValidation(
+				messageTemplate,
+				constraintViolationCreationContext.getMessageParameters(),
+				constraintViolationCreationContext.getExpressionVariables(),
+				interpolatedMessage,
+				getRootBeanClass(),
+				getRootBean(),
+				localContext.getCurrentBean(),
+				localContext.getCurrentValidatedValue(),
+				propertyPath,
+				constraintDescriptor,
+				localContext.getElementType(),
+				constraintViolationCreationContext.getDynamicPayload()
+		);
+	}
+
+	@Override
+	public String toString() {
+		final StringBuilder sb = new StringBuilder( "BeanValidationContext{" );
+		sb.append( "rootBeanClass=" ).append( getRootBeanClass() );
+		sb.append( '}' );
+		return sb.toString();
+	}
+
+}
