+++ engine/src/main/java/org/hibernate/validator/internal/engine/validationcontext/package-info.java
@@ -1,0 +1,10 @@
+/*
+ * Hibernate Validator, declare and validate application constraints
+ *
+ * License: Apache License, Version 2.0
+ * See the license.txt file in the root directory or <http://www.apache.org/licenses/LICENSE-2.0>.
+ */
+/**
+ * Contains various implementation of validation contexts and supporting classes.
+ */
+package org.hibernate.validator.internal.engine.validationcontext;
